Add support for flash TEMPPAGE writing.
The proper way to write to flash on the AVR is to write
to the temp page and then write that temp page to flash.
This allows cancellation with the RWWSRE bit for clearing
out the temp page without writing it to flash.

This fixed a bug where the RWWSRE bit was being set in a
command, but the flash driver was improperly writing whatever
was in r0/r1 to the first address in the page, overwriting
what was already there. In reality, the code running on the
core was trying to just flush the temp page (probably unnecessarily,
but was still valid code).

Unfortunately, not all AVR cores support the RWWSRE bit, so
I had to work around that using an _ext() initialization function.
Cores that support it can initialize the flash using the _ext()
function and RWWSRE support will work. Regardless, the TEMPPAGE
writing will work.